Fix integer overflow leading to undersized allocation in PFM loader (CVE-2020-22524)
Load() computed the RLE/pixel line buffer size as
`const unsigned lineWidth = 3 * width;` (RGBF) or `= width;` (FLOAT),
where width is an int read directly from the PFM header with no upper
bound (only `width <= 0` is rejected). For the RGBF case in particular,
`3 * width` is evaluated as a 32-bit signed int multiply before being
stored into lineWidth, so a sufficiently large width (~715M+) silently
overflows and wraps to a small value.

malloc(lineWidth * sizeof(float)) then succeeds with a buffer far
smaller than the real per-scanline size, while the pixel-unpacking
loop right below it still iterates `for (x = 0; x < width; x++)`
against the original, un-overflowed width - reading/writing past the
end of that undersized lineBuffer.

Compute lineWidth in size_t instead, so the multiply can't overflow
before it reaches malloc(); genuinely unreasonable widths now just
fail malloc() cleanly, which Load() already handles via the existing
`if(!lineBuffer) throw FI_MSG_ERROR_MEMORY` check.

CWE-190 (Integer Overflow) leading to CWE-787/CWE-125 (Out-of-bounds
Write/Read)

Fixes: CVE-2020-22524
(cherry picked from commit 668385e561a4a338f5b3513a6f09a0502933281e)

+++ Source/FreeImage/PluginPFM.cpp
@@ -2,7 +2,7 @@
 // PFM Loader and Writer
 //
 // Design and implementation by
-// - Hervé Drolon (drolon@infonie.fr)
+// - Hervï¿½ Drolon (drolon@infonie.fr)
 //
 // This file is part of FreeImage 3
 //
@@ -271,7 +271,13 @@ Load(FreeImageIO *io, fi_handle handle, int page, int flags, void *data) {
 		// Read the image...
 
 		if(image_type == FIT_RGBF) {
-			const unsigned lineWidth = 3 * width;
+			// Compute in size_t (not int/unsigned) so a large, attacker-
+			// controlled width can't silently overflow the 32-bit multiply
+			// before it reaches malloc()'s size argument - an overflowed,
+			// too-small lineWidth here would still let the per-pixel unpack
+			// loop below (bounded by the untruncated `width`) read past the
+			// end of this smaller-than-expected lineBuffer allocation.
+			const size_t lineWidth = (size_t)3 * (size_t)width;
 			lineBuffer = (float*)malloc(lineWidth * sizeof(float));
 			if(!lineBuffer) {
 				throw FI_MSG_ERROR_MEMORY;
@@ -305,7 +311,7 @@ Load(FreeImageIO *io, fi_handle handle, int page, int flags, void *data) {
 			lineBuffer = NULL;
 
 		} else if(image_type == FIT_FLOAT) {
-			const unsigned lineWidth = width;
+			const size_t lineWidth = (size_t)width;
 			lineBuffer = (float*)malloc(lineWidth * sizeof(float));
 			if(!lineBuffer) {
 				throw FI_MSG_ERROR_MEMORY;
